backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyjwt
bcrypt
authlib
httpx
itsdangerous
python-binance
ta
scikit-learn
matplotlib
pandas
numpy
huggingface_hub
feedparser
sqlalchemy[asyncio]
asyncpg
alembic
